Re-organized packages for consistent naming.
-Pop up window with progress bar may need multithreading processing. Until then - will use alert box with dialog for what files completed.

Stackoverflow on multithreading and progress bar - https://stackoverflow.com/questions/34773577/javafx-show-progressindicator-while-loading
