+++ src/main/java/com/poker_player_tracker/data_IO/game_processor/GameFileKey.java
@@ -1,4 +1,4 @@
-package com.poker_player_tracker.data_IO.processor;
+package com.poker_player_tracker.data_IO.game_processor;
 
 public enum GameFileKey {
